code/aml--fabric-generation-structured-data-flow/requirements.txt
azure-ai-ml
azureml-fsspec 
mltable
keyrings.alt